Move SeriesInfo to dataset module
